+++ st2actions/tests/unit/test_mistral_v2.py
@@ -33,6 +33,11 @@ from oslo_config import cfg
 import st2tests.config as tests_config
 tests_config.parse_args()
 
+# Set defaults for retry options.
+cfg.CONF.set_override('retry_exp_msec', 100, group='mistral')
+cfg.CONF.set_override('retry_exp_max_msec', 100, group='mistral')
+cfg.CONF.set_override('retry_stop_max_msec', 300, group='mistral')
+
 import st2common.bootstrap.runnersregistrar as runners_registrar
 from st2actions.handlers.mistral import MistralCallbackHandler
 from st2actions.runners.localrunner import LocalShellRunner
@@ -47,7 +52,6 @@ from st2common.services import action as action_service
 from st2common.transport.liveaction import LiveActionPublisher
 from st2common.transport.publishers import CUDPublisher
 from st2tests import DbTestCase
-from st2tests import http
 from st2tests.fixturesloader import FixturesLoader
 from tests.unit.base import MockLiveActionPublisher
 
@@ -160,11 +164,6 @@ class MistralRunnerTest(DbTestCase):
     def setUp(self):
         super(MistralRunnerTest, self).setUp()
         cfg.CONF.set_override('api_url', 'http://0.0.0.0:9101', group='auth')
-
-    def tearDown(self):
-        super(MistralRunnerTest, self).tearDown()
-        cfg.CONF.set_default('max_attempts', 2, group='mistral')
-        cfg.CONF.set_default('retry_wait', 1, group='mistral')
 
     @mock.patch.object(
         workflows.WorkflowManager, 'list',
@@ -314,14 +313,14 @@ class MistralRunnerTest(DbTestCase):
 
     @mock.patch.object(
         workflows.WorkflowManager, 'list',
-        mock.MagicMock(side_effect=requests.exceptions.ConnectionError()))
+        mock.MagicMock(side_effect=requests.exceptions.ConnectionError('Connection refused')))
     def test_launch_workflow_mistral_offline(self):
         MistralRunner.entry_point = mock.PropertyMock(return_value=WF1_YAML_FILE_PATH)
         liveaction = LiveActionDB(action=WF1_NAME, parameters=ACTION_PARAMS)
         liveaction, execution = action_service.request(liveaction)
         liveaction = LiveAction.get_by_id(str(liveaction.id))
         self.assertEqual(liveaction.status, action_constants.LIVEACTION_STATUS_FAILED)
-        self.assertIn('Failed to connect to mistral', liveaction.result['error'])
+        self.assertIn('Connection refused', liveaction.result['error'])
 
     @mock.patch.object(
         workflows.WorkflowManager, 'list',
@@ -599,23 +598,23 @@ class MistralRunnerTest(DbTestCase):
         self.assertIn('Name of the workbook must be the same', liveaction.result['error'])
 
     @mock.patch.object(
-        requests, 'request',
-        mock.MagicMock(return_value=http.FakeResponse({}, 200, 'OK')))
+        action_executions.ActionExecutionManager, 'update',
+        mock.MagicMock(return_value=None))
     def test_callback_handler_with_result_as_text(self):
         MistralCallbackHandler.callback('http://localhost:8989/v2/action_executions/12345', {},
                                         action_constants.LIVEACTION_STATUS_SUCCEEDED,
                                         '<html></html>')
 
     @mock.patch.object(
-        requests, 'request',
-        mock.MagicMock(return_value=http.FakeResponse({}, 200, 'OK')))
+        action_executions.ActionExecutionManager, 'update',
+        mock.MagicMock(return_value=None))
     def test_callback_handler_with_result_as_dict(self):
         MistralCallbackHandler.callback('http://localhost:8989/v2/action_executions/12345', {},
                                         action_constants.LIVEACTION_STATUS_SUCCEEDED, {'a': 1})
 
     @mock.patch.object(
-        requests, 'request',
-        mock.MagicMock(return_value=http.FakeResponse({}, 200, 'OK')))
+        action_executions.ActionExecutionManager, 'update',
+        mock.MagicMock(return_value=None))
     def test_callback_handler_with_result_as_json_str(self):
         MistralCallbackHandler.callback('http://localhost:8989/v2/action_executions/12345', {},
                                         action_constants.LIVEACTION_STATUS_SUCCEEDED, '{"a": 1}')
@@ -623,16 +622,16 @@ class MistralRunnerTest(DbTestCase):
                                         action_constants.LIVEACTION_STATUS_SUCCEEDED, "{'a': 1}")
 
     @mock.patch.object(
-        requests, 'request',
-        mock.MagicMock(return_value=http.FakeResponse({}, 200, 'OK')))
+        action_executions.ActionExecutionManager, 'update',
+        mock.MagicMock(return_value=None))
     def test_callback_handler_with_result_as_list(self):
         MistralCallbackHandler.callback('http://localhost:8989/v2/action_executions/12345', {},
                                         action_constants.LIVEACTION_STATUS_SUCCEEDED,
                                         ["a", "b", "c"])
 
     @mock.patch.object(
-        requests, 'request',
-        mock.MagicMock(return_value=http.FakeResponse({}, 200, 'OK')))
+        action_executions.ActionExecutionManager, 'update',
+        mock.MagicMock(return_value=None))
     def test_callback_handler_with_result_as_list_str(self):
         MistralCallbackHandler.callback('http://localhost:8989/v2/action_executions/12345', {},
                                         action_constants.LIVEACTION_STATUS_SUCCEEDED,
@@ -662,9 +661,6 @@ class MistralRunnerTest(DbTestCase):
             requests.exceptions.ConnectionError(),
             None]))
     def test_callback_retry(self):
-        cfg.CONF.set_default('max_attempts', 3, group='mistral')
-        cfg.CONF.set_default('retry_wait', 0, group='mistral')
-
         liveaction = LiveActionDB(
             action='core.local', parameters={'cmd': 'uname -a'},
             callback={
@@ -689,9 +685,6 @@ class MistralRunnerTest(DbTestCase):
             requests.exceptions.ConnectionError(),
             None]))
     def test_callback_retry_exhausted(self):
-        cfg.CONF.set_default('max_attempts', 3, group='mistral')
-        cfg.CONF.set_default('retry_wait', 0, group='mistral')
-
         liveaction = LiveActionDB(
             action='core.local', parameters={'cmd': 'uname -a'},
             callback={
@@ -742,6 +735,76 @@ class MistralRunnerTest(DbTestCase):
         executions.ExecutionManager.update.assert_called_with(WF1_EXEC.get('id'), 'PAUSED')
         liveaction = LiveAction.get_by_id(str(liveaction.id))
         self.assertEqual(liveaction.status, action_constants.LIVEACTION_STATUS_CANCELED)
+
+    @mock.patch.object(
+        workflows.WorkflowManager, 'list',
+        mock.MagicMock(return_value=[]))
+    @mock.patch.object(
+        workflows.WorkflowManager, 'get',
+        mock.MagicMock(return_value=WF1))
+    @mock.patch.object(
+        workflows.WorkflowManager, 'create',
+        mock.MagicMock(return_value=[WF1]))
+    @mock.patch.object(
+        executions.ExecutionManager, 'create',
+        mock.MagicMock(return_value=executions.Execution(None, WF1_EXEC)))
+    @mock.patch.object(
+        executions.ExecutionManager, 'update',
+        mock.MagicMock(side_effect=[requests.exceptions.ConnectionError(),
+                                    executions.Execution(None, WF1_EXEC_PAUSED)]))
+    def test_cancel_retry(self):
+        MistralRunner.entry_point = mock.PropertyMock(return_value=WF1_YAML_FILE_PATH)
+        liveaction = LiveActionDB(action=WF1_NAME, parameters=ACTION_PARAMS)
+        liveaction, execution = action_service.request(liveaction)
+        liveaction = LiveAction.get_by_id(str(liveaction.id))
+        self.assertEqual(liveaction.status, action_constants.LIVEACTION_STATUS_RUNNING)
+
+        mistral_context = liveaction.context.get('mistral', None)
+        self.assertIsNotNone(mistral_context)
+        self.assertEqual(mistral_context['execution_id'], WF1_EXEC.get('id'))
+        self.assertEqual(mistral_context['workflow_name'], WF1_EXEC.get('workflow_name'))
+
+        requester = cfg.CONF.system_user.user
+        liveaction, execution = action_service.request_cancellation(liveaction, requester)
+        executions.ExecutionManager.update.assert_called_with(WF1_EXEC.get('id'), 'PAUSED')
+        liveaction = LiveAction.get_by_id(str(liveaction.id))
+        self.assertEqual(liveaction.status, action_constants.LIVEACTION_STATUS_CANCELED)
+
+    @mock.patch.object(
+        workflows.WorkflowManager, 'list',
+        mock.MagicMock(return_value=[]))
+    @mock.patch.object(
+        workflows.WorkflowManager, 'get',
+        mock.MagicMock(return_value=WF1))
+    @mock.patch.object(
+        workflows.WorkflowManager, 'create',
+        mock.MagicMock(return_value=[WF1]))
+    @mock.patch.object(
+        executions.ExecutionManager, 'create',
+        mock.MagicMock(return_value=executions.Execution(None, WF1_EXEC)))
+    @mock.patch.object(
+        executions.ExecutionManager, 'update',
+        mock.MagicMock(side_effect=requests.exceptions.ConnectionError('Connection refused')))
+    def test_cancel_retry_exhausted(self):
+        MistralRunner.entry_point = mock.PropertyMock(return_value=WF1_YAML_FILE_PATH)
+        liveaction = LiveActionDB(action=WF1_NAME, parameters=ACTION_PARAMS)
+        liveaction, execution = action_service.request(liveaction)
+        liveaction = LiveAction.get_by_id(str(liveaction.id))
+        self.assertEqual(liveaction.status, action_constants.LIVEACTION_STATUS_RUNNING)
+
+        mistral_context = liveaction.context.get('mistral', None)
+        self.assertIsNotNone(mistral_context)
+        self.assertEqual(mistral_context['execution_id'], WF1_EXEC.get('id'))
+        self.assertEqual(mistral_context['workflow_name'], WF1_EXEC.get('workflow_name'))
+
+        requester = cfg.CONF.system_user.user
+        liveaction, execution = action_service.request_cancellation(liveaction, requester)
+
+        calls = [call(WF1_EXEC.get('id'), 'PAUSED') for i in range(0, 3)]
+        executions.ExecutionManager.update.assert_has_calls(calls)
+
+        liveaction = LiveAction.get_by_id(str(liveaction.id))
+        self.assertEqual(liveaction.status, action_constants.LIVEACTION_STATUS_CANCELING)
 
     def test_build_context(self):
         parent = {