+++ st2api/tests/integration/test_gunicorn_configs.py
@@ -24,42 +24,43 @@ from eventlet.green import subprocess
 
 from st2common.models.utils import profiling
 from st2common.util.shell import kill_process
+from st2tests.base import IntegrationTestCase
 
 
 BASE_DIR = os.path.dirname(os.path.abspath(__file__))
+ST2_CONFIG_PATH = os.path.join(BASE_DIR, '../../../conf/st2.tests.conf')
 
 
-class GunicornWSGIEntryPointTestCase(unittest2.TestCase):
+class GunicornWSGIEntryPointTestCase(IntegrationTestCase):
     @unittest2.skipIf(profiling.is_enabled(), 'Profiling is enabled')
-    @unittest2.skipIf(os.environ.get('TRAVIS'), 'Running on travis')
     def test_st2api_wsgi_entry_point(self):
         port = random.randint(10000, 30000)
         config_path = os.path.join(BASE_DIR, '../../../st2api/st2api/gunicorn_config.py')
-        st2_config_path = os.path.join(BASE_DIR, '../../../conf/st2.dev.conf')
         cmd = ('gunicorn_pecan %s -k eventlet -w 1 -b 127.0.0.1:%s' % (config_path, port))
         env = os.environ.copy()
-        env['ST2_CONFIG_PATH'] = st2_config_path
+        env['ST2_CONFIG_PATH'] = ST2_CONFIG_PATH
         process = subprocess.Popen(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, env=env,
                                    shell=True, preexec_fn=os.setsid)
-
+        self.add_process(process=process)
         eventlet.sleep(5)
+        self.assertProcessIsRunning(process=process)
         response = requests.get('http://127.0.0.1:%s/v1/actions' % (port))
         self.assertEqual(response.status_code, httplib.OK)
         kill_process(process)
 
     @unittest2.skipIf(profiling.is_enabled(), 'Profiling is enabled')
-    @unittest2.skipIf(os.environ.get('TRAVIS'), 'Running on travis')
     def test_st2auth(self):
         port = random.randint(10000, 30000)
         config_path = os.path.join(BASE_DIR, '../../../st2auth/st2auth/gunicorn_config.py')
-        st2_config_path = os.path.join(BASE_DIR, '../../../conf/st2.dev.conf')
         cmd = ('gunicorn_pecan %s -k eventlet -w 1 -b 127.0.0.1:%s' % (config_path, port))
         env = os.environ.copy()
-        env['ST2_CONFIG_PATH'] = st2_config_path
+        env['ST2_CONFIG_PATH'] = ST2_CONFIG_PATH
         process = subprocess.Popen(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, env=env,
                                    shell=True, preexec_fn=os.setsid)
 
+        self.add_process(process=process)
         eventlet.sleep(5)
+        self.assertProcessIsRunning(process=process)
         response = requests.post('http://127.0.0.1:%s/tokens' % (port))
         self.assertEqual(response.status_code, httplib.UNAUTHORIZED)
         kill_process(process)