+++ st2common/st2common/config.py
@@ -106,7 +106,7 @@ def register_opts(ignore_errors=False):
     messaging_opts = [
         # It would be nice to be able to deprecate url and completely switch to using
         # url. However, this will be a breaking change and will have impact so allowing both.
-        cfg.StrOpt('url', default='amqp://guest:guest@localhost:5672//',
+        cfg.StrOpt('url', default='amqp://guest:guest@127.0.0.1:5672//',
                    help='URL of the messaging server.'),
         cfg.ListOpt('cluster_urls', default=[],
                     help='URL of all the nodes in a messaging service cluster.')
@@ -114,7 +114,7 @@ def register_opts(ignore_errors=False):
     do_register_opts(messaging_opts, 'messaging', ignore_errors)
 
     syslog_opts = [
-        cfg.StrOpt('host', default='localhost',
+        cfg.StrOpt('host', default='127.0.0.1',
                    help='Host for the syslog server.'),
         cfg.IntOpt('port', default=514,
                    help='Port for the syslog server.'),
@@ -165,9 +165,10 @@ def register_opts(ignore_errors=False):
 
     # Mistral options
     mistral_opts = [
-        cfg.StrOpt('v2_base_url', default='http://localhost:8989/v2', help='v2 API root endpoint.'),
-        cfg.IntOpt('max_attempts', default=180, help='Max attempts to reconnect.'),
-        cfg.IntOpt('retry_wait', default=5, help='Seconds to wait before reconnecting.'),
+        cfg.StrOpt('v2_base_url', default='http://127.0.0.1:8989/v2', help='v2 API root endpoint.'),
+        cfg.IntOpt('retry_exp_msec', default=1000, help='Multiplier for the exponential backoff.'),
+        cfg.IntOpt('retry_exp_max_msec', default=300000, help='Max time for each set of backoff.'),
+        cfg.IntOpt('retry_stop_max_msec', default=600000, help='Max time to stop retrying.'),
         cfg.StrOpt('keystone_username', default=None, help='Username for authentication.'),
         cfg.StrOpt('keystone_password', default=None, help='Password for authentication.'),
         cfg.StrOpt('keystone_project_name', default=None, help='OpenStack project scope.'),