+++ st2common/st2common/models/api/base.py
@@ -24,13 +24,13 @@ import six
 from six.moves import http_client
 from webob import exc
 import pecan
-import pecan.jsonify
 import traceback
 from oslo_config import cfg
 
 from st2common.constants.pack import DEFAULT_PACK_NAME
 from st2common.util import mongoescape as util_mongodb
 from st2common.util import schema as util_schema
+from st2common.util.debugging import is_enabled as is_debugging_enabled
 from st2common.util.jsonify import json_encode
 from st2common import log as logging
 
@@ -231,7 +231,11 @@ def jsexpose(arg_types=None, body_cls=None, status_code=None, content_type='appl
             if status_code:
                 pecan.response.status = status_code
             if content_type == 'application/json':
-                return json_encode(result, indent=None)
+                if is_debugging_enabled():
+                    indent = 4
+                else:
+                    indent = None
+                return json_encode(result, indent=indent)
             else:
                 return result