+++ st2tests/st2tests/base.py
@@ -24,9 +24,10 @@ import sys
 import shutil
 import logging
 
-import eventlet
-from oslo_config import cfg
 import six
+import eventlet
+import psutil
+from oslo_config import cfg
 from unittest2 import TestCase
 
 from st2common.exceptions.db import StackStormDBObjectConflictError
@@ -54,7 +55,8 @@ __all__ = [
     'DbTestCase',
     'DbModelTestCase',
     'CleanDbTestCase',
-    'CleanFilesTestCase'
+    'CleanFilesTestCase',
+    'IntegrationTestCase'
 ]
 
 BASE_DIR = os.path.dirname(os.path.abspath(__file__))
@@ -74,6 +76,17 @@ ALL_MODELS.extend(policy_model.MODELS)
 
 BASE_DIR = os.path.dirname(os.path.abspath(__file__))
 TESTS_CONFIG_PATH = os.path.join(BASE_DIR, '../conf/st2.conf')
+
+
+class BaseTestCase(TestCase):
+
+    @classmethod
+    def _register_packs(self):
+        """
+        Register all the packs inside the fixtures directory.
+        """
+        registrar = ResourceRegistrar(use_pack_cache=False)
+        registrar.register_packs(base_dirs=get_packs_base_paths())
 
 
 class EventletTestCase(TestCase):
@@ -103,7 +116,7 @@ class EventletTestCase(TestCase):
         )
 
 
-class BaseDbTestCase(TestCase):
+class BaseDbTestCase(BaseTestCase):
 
     # Set to True to enable printing of all the log messages to the console
     DISPLAY_LOG_MESSAGES = False
@@ -148,14 +161,6 @@ class BaseDbTestCase(TestCase):
         global ALL_MODELS
         for model in ALL_MODELS:
             model.drop_collection()
-
-    @classmethod
-    def _register_packs(self):
-        """
-        Register all the packs inside the fixtures directory.
-        """
-        registrar = ResourceRegistrar(use_pack_cache=False)
-        registrar.register_packs(base_dirs=get_packs_base_paths())
 
 
 class DbTestCase(BaseDbTestCase):
@@ -316,6 +321,85 @@ class CleanFilesTestCase(TestCase):
                 shutil.rmtree(file_path)
             except Exception:
                 pass
+
+
+class IntegrationTestCase(TestCase):
+    """
+    Base test class for integration tests to inherit from.
+
+    It includes various utility functions and assert methods for working with processes.
+    """
+
+    # Set to True to print process stdout and stderr in tearDown after killing the processes
+    # which are still alive
+    print_stdout_stderr_on_teardown = False
+
+    processes = {}
+
+    def tearDown(self):
+        super(IntegrationTestCase, self).tearDown()
+
+        # Make sure we kill all the processes on teardown so they don't linger around if an
+        # exception was thrown.
+        for pid, process in self.processes.items():
+
+            try:
+                process.kill()
+            except OSError:
+                # Process already exited or similar
+                pass
+
+            if self.print_stdout_stderr_on_teardown:
+                try:
+                    stdout = process.stdout.read()
+                except:
+                    stdout = None
+
+                try:
+                    stderr = process.stderr.read()
+                except:
+                    stderr = None
+
+                print('Process "%s"' % (process.pid))
+                print('Stdout: %s' % (stdout))
+                print('Stderr: %s' % (stderr))
+
+    def add_process(self, process):
+        """
+        Add a process to the local data structure to make sure it will get killed and cleaned up on
+        tearDown.
+        """
+        self.processes[process.pid] = process
+
+    def remove_process(self, process):
+        """
+        Remove process from a local data structure.
+        """
+        if process.pid in self.processes:
+            del self.processes[process.pid]
+
+    def assertProcessIsRunning(self, process):
+        """
+        Assert that a long running process provided Process object as returned by subprocess.Popen
+        has succesfuly started and is running.
+        """
+        return_code = process.poll()
+
+        if return_code is not None:
+            stdout = process.stdout.read()
+            stderr = process.stderr.read()
+            msg = ('Process exited with code=%s.\nStdout:\n%s\n\nStderr:\n%s' %
+                   (return_code, stdout, stderr))
+            self.fail(msg)
+
+    def assertProcessExited(self, proc):
+        try:
+            status = proc.status()
+        except psutil.NoSuchProcess:
+            status = 'exited'
+
+        if status not in ['exited', 'zombie']:
+            self.fail('Process with pid "%s" is still running' % (proc.pid))
 
 
 class FakeResponse(object):