+++ tools/launchdev.sh
@@ -94,10 +94,14 @@ function st2start(){
         sudo mkdir -p $PACKS_BASE_DIR
     fi
 
+    VIRTUALENVS_DIR=$ST2_BASE_DIR/virtualenvs
+
     sudo mkdir -p $PACKS_BASE_DIR/default/sensors/
     sudo mkdir -p $PACKS_BASE_DIR/default/actions/
     sudo mkdir -p $PACKS_BASE_DIR/default/rules/
+    sudo mkdir -p $VIRTUALENVS_DIR
     sudo chown -R ${CURRENT_USER}:${CURRENT_USER_GROUP} $PACKS_BASE_DIR
+    sudo chown -R ${CURRENT_USER}:${CURRENT_USER_GROUP} $VIRTUALENVS_DIR
     cp -Rp ./contrib/core/ $PACKS_BASE_DIR
     cp -Rp ./contrib/packs/ $PACKS_BASE_DIR